Added string parameters to the groups in the params database.  This is intended be backwards compatible with existing params files; please let me know if that turns out not to be the case.
int stringPresent(groupName, parameterName) determines if a group has a string parameter.
const char * stringValue(groupName, parameterName) returns the value if the string parameter was defined,
or NULL if it wasn't.

(stringPresent isn't necessary since you can test against NULL, but I included it in analogy to present(). )
Strings can contain letters, numerals, and the underscore.  They may also contain EITHER spaces OR characters in the set {'_', '~', '+', '/', '.', '-'}, but not both.
The first possibility allows you to set strings to the names of other groups.  The second allows you to set strings to filenames.

The motivation is that all the parameters necessary to define a connection, layer, or probe should be definable within the params file.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@3618 7c0444ca-5550-0410-9008-faecdbe78eb5