TransposeConn is a derived class of PeriodicUpdateConn.  The original connection, a KernelConn, is specified in the constructor.  The presynaptic and postsynaptic layers do not have to be the postsynaptic and presynaptic layers, respectively, of the original connection, but they do have to have the same dimension.
FeedbackConn is now a special cass of TransposeConn.  The presynaptic and postsynaptic layers are not passed in the constructor; instead they are the postsynaptic and presynaptic layers, respectively, of the original connection.

tests/test_transpose_transpose.cpp now tests both TransposeConn and FeedbackConn, and has been reorganized to have less duplicated code.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@3621 7c0444ca-5550-0410-9008-faecdbe78eb5