HyPerConn::updateState calls a new method, calc_dW, before calling updateWeights.
KernelConn::updateState is similar, and calls reduceKernels between calc_dW and updateWeights.

The motivation is to separate computing the change in the weights separately from combining those changes into the weights, since for some classes it is more natural to reduce dW than the weights themselves.

KernelConn also has the buffer for MPI_Allreduce preallocated as a member variable, mpiReductionBuffer.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@3987 7c0444ca-5550-0410-9008-faecdbe78eb5