Changes to PVParams:
A duplicated parameter name in a parameter group is an error; the program will now exit if a duplicate name is encountered.

If a parameter appears in the params file but has not been queried when run is called, a warning is printed.

The FilenameDef and FilenameStack classes, which I believe have long been unused, are deprecated.  Defining one in the params file generates a warning.  These two classes will be removed soon.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4442 7c0444ca-5550-0410-9008-faecdbe78eb5