Progress toward complete checkpointing.  If the HyPerCol parameter checkpointWrite is set to true, a directory checkpointWriteDir must be defined and one (but only one) of checkpointWriteTimeInterval or checkpointWriteStepInterval must be set to a positive number; it indicates how often checkpoints are saved.  Currently, each layer writes its activity, V, GSyn and datastore to the checkpoint.  The connections' checkpointWrite still needs to be implemented.
There is a checkpointRead parameter as well, but its functionality is not yet implemented.

There is the following change to the format of the header file for PVP_NONSPIKING_ACT_FILE (filetype 4) files: the parameter INDEX_NBANDS stores the number of records in the file.  For the checkpointed A and V files, INDEX_NBANDS is 1.  For the GSyn files, INDEX_BANDS is the number of channels.  For the datastore files, it is the number of levels in the datastore.  For the output files a0.pvp, a1.pvp, etc., it is the number of timesteps that have been written.  Previously, INDEX_NBANDS had the number of features, making it redundant with INDEX_NF.


git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4469 7c0444ca-5550-0410-9008-faecdbe78eb5