Extensive changes to implement checkpointing. To write checkpoints to a directory named cpDir, set the HyPerCol parameter checkpointWrite=true and checkpointWriteDir="cpDir".  The directory doesn't have to exist but its parent directory does. This creates directories in the specified directory named Checkpoint0, Checkpoint1, etc.  The frequency of checkpoints is controlled setting either checkpointWriteStepInterval=n (to checkpoint every n-th timestep) or checkpointWriteTimeInterval=x (to checkpoint when the simulation time increases by x). To read checkpoints from cpDir, set the HyPerCol parameter checkpointRead=true and checkpointReadDir="cpDir", and specify the specific checkpoint Checkpoint[n] with checkpointReadDirIndex = [n];  If checkpointWrite is also set to true, the checkpoint index will pick up at [n+1], in the directory specified in checkpointWriteDir.  checkpointWriteDir and checkpointReadDir can be, but do not have to be, the same directory.
The way that connections with multiple arbors are read and written has been changed.  All arbors are stored in the same .pvp file.  The number of arbors is stored in the NBANDS entry in the header (bytes 68-71).  After the header, all patches with arbor 0 (across all processes) are stored, then all patches with arbor 1, and so forth.

If initializing weights using FileWeight, a connection with several arbors can be specified with different arbors in different files.  Create a text file where each line has one of the .pvp files and use the filename of the text file as the argument for the initWeightsFile parameter.  Then set the parameter useListOfArborFiles = true

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4586 7c0444ca-5550-0410-9008-faecdbe78eb5