IncrementLayer is a derived layer of ANNLayer.  A framerate is specified with the params file parameter displayPeriod and a starting time is specified with the parameter firstUpdateTime.
When t=firstUpdateTime, the excitatory channel is copied to V.  After an additional time of displayPeriod has elapsed, V is moved to Vprev, the excitatory channel is copied to V, and the activity buffer contains V-Vprev.  Hence the activity buffer contains the amount by which the excitatory channel incremented over an interval of displayPeriod.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4759 7c0444ca-5550-0410-9008-faecdbe78eb5