An overhaul of the Patterns class.  The change in behavior should be limited to the BARS patternType.  In the params file, width and height set the thickness of the bars when in their vertical and horizontal orientations, respectively.  wavelengthVert and wavelengthHoriz set the period of the grating pattern; the period is no longer limited to twice the thickness.  If width is an integer n, the thickness that results when position is an integer is now n, not n-1.  The code assumes but doesn't check that the thickness is at least one pixel less than the wavelength.
Before, moving and switching orientations were governed by independent random variables, with the result that the probability of each happening was not what one would naturally expect.  Now, a single random number in U(0,1) is generated.  If it's in the interval (0,pSwitch), the orientation switches.  If it's in the interval (pSwitch,pSwitch+pMove), the orientation stays the same.  In either event, the position is recalculated using the movementType specified in the params.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4762 7c0444ca-5550-0410-9008-faecdbe78eb5