Overhaul of updateActiveIndices.
updateActiveIndices will not be put on the GPU.  Hence it is called by the public updateState(float time, float dt) method.  HyPerLayer's updateActiveIndices method checks spikingFlag and, if true, calls calcActiveIndices.  It's virtual and can be overridden.  The only class that overrides it right now is CliqueLayer, which calls calcActiveIndices without checking spikingFlag.  This is equivalent to the previous behavior but eliminates code duplication between CliqueLayer's and HyPerLayer's updateActiveIndices.

This update passes all unit and systems test with PV_USE_MPI defined and either with or without PV_USE_OPENCL

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4915 7c0444ca-5550-0410-9008-faecdbe78eb5