Overhaul of LayerProbe.  LayerProbe constructors take a HyPerLayer as an argument, and LayerProbe::outputState no longer takes a HyPerLayer as an argument.  Hence every layer probe belongs to a specific layer.  When a HyPerLayer's insertProbe method is called, it checks that the probe has the correct target layer before adding it to the list of probes.  It also checks that the probe is not already in the list of probes and won't add the same probe twice.  It issues warnings if either of these checks fails.
The constructors for the layer probe classes call an initialization function usually called init<ClassName>.  I say usually because I tried to be consistent but might have missed here and there.  For the same reasons we did this in HyPerLayer and HyPerConn, I had the derived class constructor call the default constructor of the base class, and the derived class init method calls the base class init method.

The initLayerProbe method calls HyPerLayer::insertProbe.  Therefore, the function that creates the LayerProbe should not call insertProbe.  (However, as mentioned above the same layer will not be inserted twice, so doing so should have no effect except for an avoidable warning being produced.)

I also made some cosmetic changes along the way, principally changing the name of the float time argument to outputState and other routines to timef.  The change was because time is the name of a built-in library function; the appended f is because the time is passed as a float.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4917 7c0444ca-5550-0410-9008-faecdbe78eb5