A mostly cosmetic change to the way StatsProbe and PointProbe handle the message (the prefix to strings printed by outputState that identify the probe).  Before, the default message would be the layer's name (padded with spaces to length 31) followed by a colon.  A message could be specified in the params file, but the parser does not accept colons in strings.
The new behavior is that the msg member consists of the message parameter from the params file, followed by a colon.  The default message, used if message does not appear in the params file, does not contain a colon.

If the message passed to the StatsProbe or PointProbe constructor is empty or null, the msg member variable is the empty string (with no colon).

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4976 7c0444ca-5550-0410-9008-faecdbe78eb5