Changes to the way image files are read.  The changes should be backward compatible for image file formats and compressed pvp image files.
The main change is that uncompressed pvp files can be used as the input to Image::readImage().  A consequence of this is that gather routines in imageio.cpp now have the buf argument as float instead of unsigned char.  These changes then required changes in tiff.{c,h}, InitV.cpp and the ImageCreator class.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5029 7c0444ca-5550-0410-9008-faecdbe78eb5