If an image understood by GDAL (png, jpg, tif, etc.) is loaded and converted to grayscale, petavision now checks if the image is grayscale+alpha, rgb, or rgb+alpha.  If there is an alpha channel, it is discarded.  If the image is RGB, the conversion to grayscale uses weights 0.30, 0.59, 0.11.
Images loaded via PVP, or images not converted to grayscale, are handled the same as before.

It may be necessary to add the directory containing gdal.h (/opt/local/include for macs that installed gdal through macports) to the Includes directories in projects that use the petavision library, even if they didn't need it before.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5254 7c0444ca-5550-0410-9008-faecdbe78eb5