initializeV() in HyPerLayer, Image and Retina is now called initializeState().  In HyPerLayer, it initializes the V buffer as before, and then calls the new virtual function setActivity().  setActivity calls setActivity_HyPerLayer; the subclasses GenerativeLayer, IncrementLayer, and GapLayer override setActivity with calls to their own setActivity_* methods.  Image's and Retina's initializeState do not initialize A; they eliminate the V buffer as before.
The reason for this change is that without initializing the activity buffer, initializeV had essentially no effect.  On the first timestep, the activity was zero, and by the next timestep when A was updated, the previous value of V had already been clobbered.

This fix exposes that ArborSystemTest, GenerativeConnTest and GPUSystemTest have been broken for a while.  The params files for those tests will be updated in the next update.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5275 7c0444ca-5550-0410-9008-faecdbe78eb5