A reworking of checkpointing.  Externally, the files read from and written to are the same, but internally the implementation is different.
HyPerCol's, HyPerLayer's, and HyPerConn's checkpointRead and checkpointWrite now take a directory as an argument.  Files in the checkpoint are read to/written from that directory.  If checkpoint writing is on, HyPerCol::run and HyPerCol::exitRunLoop determine (and create, if necessary) the directory, using the timestep number, and then call HyPerCol's checkpointWrite. If checkpoint reading is on, HyPerCol::run determines the directory from the parameters file, and calls HyPerCol's checkpointRead.  (Before, HyPerCol's checkpointR/W would determine the directory, change directories, call HyPerConn's and HyPerLayer's checkpointR/W routines, and then cd back to the working directory.)

The new procedure is preferable because the working directory is always the same, and because checkpoints can be read to or written from outside of the currently existing loop.  For example, it would be better to implement the saving of data on exit, currently done as _last.pvp files, as a chekpoint into a Last directory.  Currently, the _last.pvp files do not have enough information to completely restore the state of layers and connections on exit.
I plan to add this capability soon.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5320 7c0444ca-5550-0410-9008-faecdbe78eb5