Add parameter sweeps to the params parser.
Implementing this required changes to buildandrun, HyPerCol, and InterColComm, but everything should be backwards compatible.

Adding a section to the params file (outside of connections, layers and probes) like the following:

ParameterSweep "group name":parameterName = { 1.0; 2.0; 3.0; 4.0; }; // note the semicolons after each value and after the closing brace

and running this params file from build and run will cause four runs to be executed.  In the first, the HyPerCol/layer/connection/probe named "group name" has parameterName set to the value 1.  In the second, it has the value 2, and so on.

You can have several ParameterSweeps, but all must have the same number of values.  You can have ParameterSweeps for string parameters as well:

ParameterSweep "group name":parameterString = { "string_a"; "string_b"; "string_c", "string_d"; };

Each run in the parameter sweep gets written to a different outputPath.  You can specify the outputPath directly with a 'ParameterSweep "column":outputPath' statement; or if this is absent, the runs will be saved to "output0", "output1/", etc.  (if there are more than ten values, leading zeros are added so that all output directory names have the same length and they sort numerically in directory listings).



git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5469 7c0444ca-5550-0410-9008-faecdbe78eb5