Changes to SigmoidLayer, LIFGap, and Retina to correct bugs and improve robustness for different values of dt.
Retina is changed so that the firing rate is independent of dt: if the retina fires once every 10 timesteps with dt=1, it will fire once every 20 timesteps with dt=0.5.  Remember that the parameters of Retina are measured in milliseconds, not timesteps.

SigmoidLayer has its activity multiplied by dt, so that its output is the probability of a spike in a timestep instead of probability per millisecond

LIFGap has a new, method for numerically integrating the differential equation for V, using Heun's method.  The string parameter "method" can be set to either "original" or "beginning".  The default is "original"; it uses what's been on the repository until now (so this change is backwards-compatible).  "beginning" uses Heun's method with the input applied at the beginning of the timestep.


git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5498 7c0444ca-5550-0410-9008-faecdbe78eb5