Changes to InitWeights and a few subclasses: If the patches argument to InitWeights::initializeWeights is null, it is assumed that the calling conn is a KernelConn.  In this case, only the root process calls calcWeights, and then the root process's weights are sent to the other processes via MPI_Bcast.
Some of the subclasses were initializing the delays for multiple arbors.  Delays for multiple arbors can now be specified in the params file, so this behavior has been removed from the subclasses.  If calcWeights were allowed to change the delays, we would have to synchronize the delays across all arbors as well.

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@6356 7c0444ca-5550-0410-9008-faecdbe78eb5