Nonspiking activity layers are written in an MPI-independent way; all files have nxprocs=1 and nyprocs=1 in the header.  The routines that read nonspiking activity layers require that nxprocs and nyprocs both be 1.  (For now, weight files are still saved in an mpi-dependent form, but this should change before long.)
The HyPerLayer parameter spikingFlag is deprecated in favor of the name writeNonspikingActivity, and the parameter writeNonspikingActivity is deprecated in favor of setting writeStep to a negative value.  Warnings are produced if the old parameters are used.

GDALOpen() calls are now handled by PV_GDALOpen, which tries GDALOpen up to five times, pausing one second in between unsuccessfulcalls, before failing.  This is an attempt to work around some apparent NFS problems on the server. 

git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@6468 7c0444ca-5550-0410-9008-faecdbe78eb5