New layer PtwiseLinearTransferLayer
PtwiseLinearTransferLayer creates a nonspiking layer with a pointwise linear
transfer function.  Array parameters verticesV and verticesV define the
vertices of the graph of the transfer function, and slopeNegInf and slopePosInf
define the slopes before the first vertex and after the second vertex.

ANNLayer is now defined as a derived class of PtwiseLinearTransferLayer.