New HyPerLayer parameters triggerBehavior and triggerResetLayerName
triggerFlag is deprecated; setting triggerLayerName to "" or NULL
implies triggerFlag is false; setting triggerLayerName to a nonempty
string implies triggerFlag is true.

The new parameters have been added to ioParamFillGroup, but currently
have no effect.  The behavior will be as follows:

triggerBehavior has two possible values:
   updateOnlyOnTrigger or resetStateOnTrigger
   The default is updateOnlyOnTrigger.

updateOnlyOnTrigger will behave as triggering has done up to now:
recvSynapticInput() and updateState() are called only when the trigger occurs.

resetStateOnTrigger will call updateState() on timesteps where the
trigger *doesn't* occur.  When the trigger does occur, it copies the
membrane potential from the layer specified in triggerResetLayerName, and
then calls setActivity().

triggerResetLayerName is read if triggerBehavior is set to resetStateOnTrigger.
If it is null or empty, it is taken to be the same as triggerLayerName.