Adds getValues() and getValue() interface to probes.
Instead of values being computed inside outputState only for the purpose
of writing to a text file, probes have getValues() and getValue() methods
that outputState calls.  The motivation is to allow a program to query
a probe for control purposes.  For example, if the HyPerCol sets
dtAdaptFlag to true and dtAdaptController to a ColProbe,
the HyPerCol uses the ColProbe to set timeScaleTrue.

Typically, probes have a number of values equal to the parent HyPerCol's
batch size.  Thus, for example, L2NormProbe computes the L2Norm probe
of each element in the batch.  However, some probes specify a batch
element as a parameter, and numValues is the number of quantities that
probe computes.  For example, PointProbe has getNumValues() returning 2,
for the membrane potential and the activity of the probed point.

Some probes that calculate large numbers of values (e.g. StatsProbe,
KernelProbe, L2ConnProbe) do not use this interface (yet).  Instead,
getNumValues() returns -1 and outputState directly computes the values
it outputs as before.