Adds get-methods numberOfBaseProbes() and getBaseProbe(int) to HyPerCol.
The baseProbes array is set during initialization, but the probes
are not handed off to layers/connections or added to the colProbes array
until HyPerCol::run() is called.
These get-methods allow inspection of the probes after build() is called
but before HyPerCol::run(), which is the most convenient place to
do such an inspection.