pv-core/CMakeLists.txt checks CUDNN_PATH
CMake exits with a descriptive error message if CUDNN_PATH is not a directory.
It will also search CUDNN_PATH recursively for the header file cudnn.h and the
library file cudnn.dylib (OS X) or cudnn.so (Linux).   If it does not find
exactly one cudnn.h and exactly one library file, it will exit with an error.