Moved parameters for the demo into the params file
Changed HeatMapProbe to read imageLayer, resultLayer, and classNames params from a params file instead of the config file. This allows for the compiled demo to be called with different networks without having to supply a different config file. The template-config and older parameter files in the PASCAL demo were updated to reflect this change. Also, the confidence table path was hard coded? This is now a parameter instead.