Fixed adaptive timestep bug in demo params files
Adding a movie layer to trigger computations needed for intermediate analysis (for the web demo) was causing the adaptive timestep to reset to dtScaleMin on every flip of the movie layer. This is needed because ImageFromMemoryBuffer can't be used to trigger other layers. Instead the adaptive timestep (in the two params files) is now calculated using probes.