Rearranges heatMapMontage.
Commands only needed to create the montage are put toward the end.
Commands that are used to create the confidence scores are moved to the start.
This is to help in moving computation of the confidences out of octave and
into C++, so that octave is only necessary for producing the image itself,
not the data used to create it.

Additionally, the square aspect-ratio params are updated to use UpdateTrigger
as the trigger layer instead of Image, and the README file is updated.