Added setOutputFilenameBase to LocalizationProbe
setOutputFilenameBase will take a path, remove the directory and extension, and set outputFilenameBase to the remaining string. The default behavior of LocalizationProbe is maintained. Output files will be prepended with "out" if setOutputFilenameBase is not called with a path string. This is so that the harness (which doesn't specify a filename) will still work, and the input name can be prepended to output filenames for use in the web application.