LocalizationProbe.hpp defines a struct "localization" to hold a detection's feature number, bounding box, and confidence.
The probeValues buffer of a LocalizationProbe now only contains a single value, the number of detections.
To retrieve the detection of a specific index, call LocalizationProbe::getDetection(index).
To retrieve a vector of all detections, call LocalizationProbe::getDetections().
The bounding box information is in the coordinates of the imageLayer, not the targetLayer.

This update also corrects confusion between the names detectionThreshold/detectionThresholds and between
heatMapMaximum/heatMapMaxima that was introduced when these parameters were changed to arrays.  In both cases,
all references have been changed to the singular version of the name.