Fixed fontsize and the way progress is printed.
The fonts for all of the labels in the heat map montage are now 18 to accommodate a smaller crop input. The progress is now "t = simtime/stoptime (percentage)" instead of "t = simtime, finish time = stoptime (percentage)", and it doesn't replace the percentage with "completed" when simtime==stoptime. This is for the video case, where HyPerCol->run is called for each video frame. Displaying finished is confusing, so now the progress display indicates the progress for each frame. The frontend of the web demo can append an indicator that the run is finished, for either the single frame or video case.