The macros pvInfo etc. take an optional argument, for assigning to a variable.
For example, the statement "pvError(errorMessage);" declares
the variable errorMessage as a _Log<char,ErrorLogType>.
This allows the printing of an error message to be split
among several print statements.  For example,
    pvError().printf("Error message line 1\n");
    pvError().printf("Error message line 2\n");
wouldn't work because the ErrorLogType kills the run after line 1.
Instead, use
    pvError(errorMessage);
    errorMessage.printf("Error message line 1\n");
    errorMessage.printf("Error message line 2\n");
Note that the prefix "ERORR <file:line>: " only gets printed once.

As an illustration, if HyPerCol::initialize gets an error when trying to
set the working directory (command line -w option), the error message
is now printed using this version of pvError instead of
two fprintf(stderr,...) statements followed by an exit statement.