Addresses the issue in <https://github.com/PetaVision/OpenPV/issues/51>.
There is a new class, normalizeGroup, that is created when
normalizeMethod="normalizeGroup".  Functionality that set up a normalizer
group has been moved from HyPerConn into this class.  The keyword
"normalizeGroup" is registered to this class, eliminating the error
message that was caused when normalizeMethod was set to "normalizeGroup".

If the normalizeMethod is set to none, there is no longer a call to
PV_Init::create made with that keyword, eliminating the resulting error.

Additionally, the HyPerConn constructor that takes an InitWeights
and NormalizeBase object has been deprecated in favor of one that
only takes the name and HyPerCol object (consistent with constructors
for other types of objects).  The createHyPerConn function has been
updated to use the new constructor.