Runs params files ending with .params.lua as a lua program.
PetaVision runs the lua program using the Lua C API, and then looks
for a lua string called "paramsFileString".  This string is then parsed
the same way that the contents of an ordinary params file is parsed.

PVModule has a new function, createParamsFileString, that returns a
string for the entire params file.  The .params.lua file generated
by HyPerCol::outputParams calls this function and assigns the result
to paramsFileString, so that it can be used directly as the params file
of another run.