Octave utilities for converting image files to pvp files
imagetopvp takes a single image file and converts it to
a pvp nonsparse activity file with a single frame.

imagelisttopvp takes a list of image files (such as the
inputPath file of the Movie layer) and converts it to
a pvp nonsparse activity file with one frame per line
of the image list (i.e. it can be the inputPath file
of a MoviePvp layer).