DataStore buffers are vectors of floats instead of char*
Several DataStore data members have been converted
from C-style arrays to vectors of size numLevels.
However, the routines that use the DataStore still
expect that all batch elements are allocated
in a single block; accordingly, mBuffer[level]
is a vector<float> of size numBuffers*numItems,
instead of a vector of vectors.

Layers and connections do not have to cast the
results of DataStore->buffer() to float.

PoolingIndexLayer stores its integral values as
floats.  Casting between int* and float* has been
eliminated; instead casting to float is done
before storing in the buffer, and casting to int
is done after retrieving from the buffer.