Secretary::checkpointWrite is called each timestep.
It uses the HyPerCol parameters related to
checkpointing to decide whether to write the
checkpoint.  It also generates the checkpoint
directory based on the step number.

Currently, HyPerCol still reads and writes the
checkpointing parameters, and Secretary
redundantly reads them.  The next update
will move the responsibility for reading and
writing to Secretary, and HyPerCol will
call get-methods as needed (which hopefully
won't be too often).