initializeFromCheckpointDir is obsolete.
Related params defaultInitializeFrmCheckpointFlag
and initializeFromCheckpointDir have been removed.
If initializeFromCheckpointDir is set to the
empty string or to NULL, a warning is generated.
If it is set to a nonempty string, it exits
with an error message.

ConnectionRestartTest and LayerRestartTest used
this feature, so they have been removed.
CommandLineRestartTest has been added, to test
restarting using the -c command line option.