Message types for auxiliary checkpointing routines
The Secretary class defines two message types that
it sends at the end of checkpointRead or the start
of checkpointToDirectory. They provide a way for
objects to get data synced to the checkpointed
buffer, if necessary.

For example, when using CUDA, MomentumLCALayer
keeps its prevDrive buffer on the GPU and only
updates it when necessary. It will implement
prepareCheckpointWrite/processCheckpointRead
by copying prevDrive from/to the GPU.