Adds ConfigParser class, for using config files
PV_Arguments has been replaced by CommandLineArguments.
There is a new class, ConfigFileArguments.  Both of the above
classes are subclasses of Arguments.  These three classes
use ConfigParser to extract the params file etc.

ConfigFileArguments allows the configuration parameters to be specified
in a configuration file instead of the command line. If the first
argument on the command line does not begin with a hyphen, PV_Init
interprets it as a config file and uses ConfigFileArguments.
If the first argument begins with a hyphen, the argc and argv
are passed to CommandLineArguments.

The rationale for ConfigFileArguments is that as the number
of configuration arguments grows, it becomes unwieldly to fit
them all on the command line, and to choose options to label them.