Configuration values retrieved by string name
Instead of get-methods such as getRequireReturn, getOutputPath, etc.,
The Configuration class and classes built on top of it now have
getBooleanArgument, getIntegerArgument, etc. that take a string
parameter. For example, getOutputPath is replaced by
getStringArgument("OutputPath"), and setOutputPath("output/") is
replaced by setStringArgument("OutputPath", "output/").

NumThreads, for which a default option is possible, is retrieved
using getIntOptionalArgument("NumThreads"). It returns a
Configuration::IntOptional struct, which has two fields, mUseDefault
and mValue.

The rationale for this commit is that as new config settings are added,
they only have to be added to the Configuration constructor, and those
places in the code where the new setting is used.