Reorganizes initialization of objects' state
When initializing the state of the objects in the HyPerCol, the previous
behavior of HyPerCol was to send InitializeStateMessage with a Checkpointer
argument. The objects responding to InitializeStateMessage would be
expected to handle InitializeFromCheckpointFlag at this point.

Then, if the CheckpointReadFlag was set (either by -c or -r on the
command line), HyPerCol would call Checkpointer::checkpointRead.

The new behavior for HyPerCol is:
Send an InitializeStateMessage, which now has no arguments.
If CheckpointReadFlag is set, call Checkpointer::checkpointRead;
otherwise send a ReadStateFromCheckpointMessage, which takes a
Checkpointer argument.
Then, send a CopyInitialStateToGPUMessage. Layers using the GPU
call copyInitialStateToGPU in response to this message.

readStateFromCheckpoint has been added to CheckpointerDataInterface.
Each class that overrides readStateFromCheckpoint checks its own
initializeFromCheckpointFlag and only calls the
Checkpointer::readNamedCheckpointEntry method if the flag is true.