Publisher's publish method split into two methods.
Call publish() if there is new data to be put into the top level of
the DataStore. Call copyForward() if the data has not changed (so that
the previous level in the DataStore needs to be copied to the top).

HyPerLayer sets a flag in updateState, and uses the flag to decide
which of the two methods to call during LayerPublish.