{read,write}NonsharedWeights uses border region.
The fields nxGlobal and nyGlobal in the pvp header have been renamed
nxExtended and nyExtended. For layers and shared-weight connections,
they continue to have the same value as nx and ny. For nonshared-weight
connections, they indicate the size of the extended region of the
pre-synaptic region required by the connection. (The presynaptic layer
may in fact be larger, due to other requirements on its border region,
but the excess is not written from or read into.)

This change breaks HyPerConn::writePostSynapticWeights, and the unit
test test_post_weights, which assumes that the post-weights PVP file
has no border region. Since writePostSynapticWeights is not being
used, and one could create a TransposeConn to get weights from the
post-synaptic perspective. Therefore, writePostSynapticWeights and
test_post_weights have been removed.

There are new unit tests, test_createPatchListDescription and
test_calcMinMaxPatch, that test the functions used to manage which
patch indices need to be read/written, and to compute the min/max
weights over several patches.

The system tests based on CheckpointSystemTest were fundamentally
broken, as they were not calling the routine that compared checkpoints.

This commit also makes other small bugfixes and improvements.