Overhauls the output stream management of probes
For LayerProbes and ColProbes, the processes with MPIBlock row index=0
and MPIBlock column index=0 open a vector of PrintStreams whose length
is the number of batch elements per process. If probeOutputFile is set,
the global batch index is inserted into the filename. Processes with
row index nonzero or column index nonzero have an empty vector of
PrintStreams. The outputState method should communicate each batch
element's information to the process with row=0,column=0 and the
appropriate batch index.

An exception is PointProbe, where the processes holding the specified
neuron and batch index have a vector of PrintStreams of length one,
and other processes have an empty PrintStreams vector.

For probes derived from BaseConnectionProbe, the root process of
each MPIBlock has a vector of PrintStreams of length one; other
processes have an empty vector. This is on the assumption that
connections use the same weights for each batch element.