Weights class, to encapsulate HyPerConn weights
This class will replace several HyPerConn data members:
wPatches, gSynPatchStart, and aPostOffset will be implemented as a
PatchGeometry data member of a Weights object; and the wDataStart
array will be implemented as the mData vector.

The dwDataStart and (in MomentumConn) the prev_dwDataStart array will
be implemented in additional Weights objects.

The PatchGeometry data member is implemented as a shared pointer,
so that these Weights objects will be able to share the same
PatchGeometry object.