normalizePixels fix, and new InputRegionLayer
Fixes an issue with InputLayer::normalizePixels, which would include
the pad region when calculating the luminance normalization,
and would normalize in the pad region.

Internally, the InputLayer calculates an extended buffer, whose values
are nonzero where the values are loaded from an image, and zero in
any pad regions.

The new layer InputRegionLayer does not have a GSyn or V buffer, and
its activity buffer points into a specified InputLayer. This way,
a mask consisting only of the region occupied by the actual image.

A new test directory, InputRegionLayerTest, is included, although
so far, this test only prints the contents of an InputLayer and an
InputRegionLayer to InfoLog; it does not yet try to detect whether
anything is incorrect.