Fixes for BinningLayer and BinningLayerTest
BinningLayer parameter normalDist has been eliminated; if binSigma is
positive, the Gaussian is always normalized to have a maximum possible
value of 1. The values depend not just on which bin the input value
falls in, but where in the bin it falls. Additionally, if binSigma > 0,
values are computed for all bins, not just those within binSigma bins
of the center.

BinningLayerTest now tests positive binSigma case correctly.