prevActivity field removed from clayer
Retina was the only class using prevActivity. It now has a new data
member, float *mSinceLastSpike. This holds the elapsed time since the
last spike, as opposed to the simulation time when that spike occurred.
The advantage is that it is not necessary to compute the difference
simTime - prevActivity, which could lead to a loss of precision.