Improves handling of initializeFromCheckpointFlag
The data member mInitializeFromCheckpointFlag is defined only once,
in CheckpointerDataInterface, with a default value of false. Derived
classes that need to initialize from checkpoint must implement a way
to set this flag.

BatchIndexer, as before, sets the flag during instantiation.

ParamsInterface has an ioParam_initializeFromCheckpointFlag function.
This method will always read the flag from params, but will only write
to the output params file if a second flag is true. Currently layers
and connections write the flag, but components within those objects
do not. This way, the flag gets written only once per params group,
as it should be.

Finally, the readStateFromCheckpoint method only gets called if
the initializeFromCheckpointFlag is true; implementations do not
need to check the flag before proceeding.