Layer updateState now handled by ActivityComponent
The ActivityComponent itself is typically made up of several
ComponentBuffer objects, with V, A, conductances, etc.
instances of ComponentBuffer subclasses.

HyPerLayer->updateState now calls the ActivityComponent's
updateActivity method, which typically calls the components'
updateBuffer method. ComponentBuffer::updateBuffer then
calls either updateBufferCPU or updateBufferGPU, which are
each virtual. However, there are some special cases, such as
LIFActivityComponent, where the interaction of the components'
update rules make it easier of for the ActivityComponent to
handle the update itself.