GSynAccumulator component
This component combines the channels of the LayerInputBuffer into
a single channel, to be used by the InternalStateBuffer.
Array parameters channelIndices and channelCoefficients specify
the multiplier for each GSyn channel.

The connection parameter convertRateToSpikeCount parameter has been
removed; its effect can be duplicated where needed by setting
the channelCoefficient in the postsynaptic layer.